--- applicaiton.js
@@ -1,47 +1,0 @@
-var express = require("express")
-var app = express()
-var bodyParser = require("body-parser")
-app.set("view engine", "hbs")
-app.use(bodyParser.json()) //handles json post requests
-app.use(bodyParser.urlencoded({ extended: true })) // handles form submissions
-
-
-
-var tasks = [
-{ id: 1,
-  body: "Blockbuster",
-  completed: false
-},
-{ id: 2,
-  body: "Roy Rogers",
-  completed: false
-},
-{ id: 3,
-  body: "Tower Records",
-  completed: false
-},
-{ id: 4,
-  body: "McDonalds",
-  completed: false
-},
-{ id: 5,
-  body: "Chucky Cheese",
-  completed: false
-}
-];
-
-app.get("/", function(req, res){
-  res.json(tasks)
-});
-
-app.get("/todo/:id", function(req,res){
-  res.json(tasks[req.params.id])
-});
-
-app.post("/", function(req,res){
-  res.json({id:6,body:'Taco Bell',completed: false})
-})
-
-app.listen(4000, function(){
-  console.log("app listening on port 4000")
-})
